Switching to stale LinkedIn profile since ProxyCurl is super expensive
Removed spending on categories for company profile as well. Updated UI to not render the column anymore.

We will spend $60-120 per month at this rate on them.

Have to cut costs and be ok with some stale data.

If it is really bad or any customer complains, then we can switch back.